bugs fixed

+++ ft_printf_dir/len_of_chars.c
@@ -6,7 +6,7 @@
 /*   By: ydag <ydag@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/11/30 13:26:21 by ydag           #+#    #+#                */
-/*   Updated: 2019/11/30 13:29:05 by ydag          ########   odam.nl         */
+/*   Updated: 2019/11/30 16:36:29 by ydag          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ int		len_of_chars(t_check chk_frmt, int num)
 		else
 			len = chk_frmt.field_width;
 	}
-	if (chk_frmt.field_width <= ft_strlen(ft_itoa(num)) && num >= 0)
+	if (chk_frmt.field_width <= (int)ft_strlen(ft_itoa(num)) && num >= 0)
 		len++;
 	return (len);
 }
